refactor: Update import statements in test files to use unified model import
