+++ src/model/__init__.py
@@ -1,0 +1,4 @@
+from .CatBoostTrainer import CatBoostTrainer
+from .LightGBMTrainer import LightGBMTrainer
+from .RandomForestTrainer import RandomForestTrainer
+from .XGBoostTrainer import XGBoostTrainer
